requirements.txt
fastapi==0.96.0
tortoise-orm==0.19.3
email-validator==2.0.0.post2
